pipeline/pyproject.toml
[project]
name = "scarti"
version = "0.1.0"
description = "Pipeline di data journalism per anomalie statistiche dell'economia italiana"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pandas>=2.2",
    "numpy>=2.0",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
]

[project.optional-dependencies]
detect = ["statsmodels>=0.14", "scipy>=1.13"]
narrate = ["anthropic>=0.40"]
mail = ["resend>=2.0"]
subscribers = ["supabase>=2.5"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21", "ruff>=0.6", "mypy>=1.11"]

[project.scripts]
scarti = "scarti.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scarti"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
